% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{download_file}
\alias{download_file}
\title{Download a csv file specified in the body parameter}
\usage{
download_file(url, file, body, hdr = headers)
}
\arguments{
\item{url}{endpoint for the download}

\item{file}{name of the file to be saved}

\item{hdr}{HTTP headers for additional information}

\item{bod}{contains data to be sent to the download endpoint}
}
\value{
download status
}
\description{
This function takes three parameters
The download endpoint, name of the file to be saved and the details for the file being requested
}
