% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{process_download}
\alias{process_download}
\title{Process download}
\usage{
process_download(
  fileName,
  fileFolder = "",
  user = Sys.getenv("ONA_USER"),
  pass = Sys.getenv("ONA_PASS"),
  base_url = Sys.getenv("BASE_URL"),
  end_point = "/v1/ona-data/download",
  auth_type = "basic",
  .headers = list()
)
}
\arguments{
\item{fileName}{Name of the file being downloaded}

\item{fileFolder}{if file exists within a sub folder provide the sub-folder path}

\item{user}{username for accessing the api if empty it will be looked for in the ONA_USER env variable}

\item{pass}{password for accessing the api if empty it will be looked for in the ONA_PASS env variable}

\item{base_url}{domain name for the api if empty it will be looked for in the BASE_URL env variable}

\item{end_point}{resource where the file will be served}

\item{auth_type}{authentication type default is basic}
}
\value{
download status fo the file
}
\description{
This function takes parameters and process the download for the file being requested
}
